# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AugmentedQuill",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_name}/sourcebook/keywords": {
      "post": {
        "tags": ["Sourcebook"],
        "summary": "Generate Sourcebook Keywords",
        "description": "Generate keywords from an entry description without persisting the entry.",
        "operationId": "generate_sourcebook_keywords_api_v1_projects__project_name__sourcebook_keywords_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourcebookKeywordsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourcebookKeywordsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_name}/sourcebook": {
      "get": {
        "tags": ["Sourcebook"],
        "summary": "Get Sourcebook",
        "description": "Return sourcebook.",
        "operationId": "get_sourcebook_api_v1_projects__project_name__sourcebook_get",
        "parameters": [
          {
            "name": "project_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Directory name of the project",
              "title": "Project Name"
            },
            "description": "Directory name of the project"
          },
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          },
          {
            "name": "match_mode",
            "in": "query",
            "required": false,
            "schema": {
              "enum": ["direct", "extensive"],
              "type": "string",
              "default": "extensive",
              "title": "Match Mode"
            }
          },
          {
            "name": "split_query_fallback",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Split Query Fallback"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SourcebookEntry"
                  },
                  "title": "Response Get Sourcebook Api V1 Projects  Project Name  Sourcebook Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Sourcebook"],
        "summary": "Create Sourcebook Entry",
        "description": "Create Sourcebook Entry.",
        "operationId": "create_sourcebook_entry_api_v1_projects__project_name__sourcebook_post",
        "parameters": [
          {
            "name": "project_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Directory name of the project",
              "title": "Project Name"
            },
            "description": "Directory name of the project"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourcebookEntryCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourcebookEntry"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_name}/sourcebook/{entry_name}": {
      "put": {
        "tags": ["Sourcebook"],
        "summary": "Update Sourcebook Entry",
        "description": "Update Sourcebook Entry.",
        "operationId": "update_sourcebook_entry_api_v1_projects__project_name__sourcebook__entry_name__put",
        "parameters": [
          {
            "name": "entry_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entry Name"
            }
          },
          {
            "name": "project_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Directory name of the project",
              "title": "Project Name"
            },
            "description": "Directory name of the project"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourcebookEntryUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourcebookEntry"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Sourcebook"],
        "summary": "Delete Sourcebook Entry",
        "description": "Delete Sourcebook Entry.",
        "operationId": "delete_sourcebook_entry_api_v1_projects__project_name__sourcebook__entry_name__delete",
        "parameters": [
          {
            "name": "entry_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entry Name"
            }
          },
          {
            "name": "project_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Directory name of the project",
              "title": "Project Name"
            },
            "description": "Directory name of the project"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Delete Sourcebook Entry Api V1 Projects  Project Name  Sourcebook  Entry Name  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_name}/search": {
      "post": {
        "tags": ["Search"],
        "summary": "Search Project",
        "description": "Search across the active project.\n\nSupported scopes: current_chapter, all_chapters, sourcebook, metadata, all.\nSupported modes: literal (default), regex (is_regex=true), phonetic (is_phonetic=true).",
        "operationId": "search_project_api_v1_projects__project_name__search_post",
        "parameters": [
          {
            "name": "project_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Directory name of the project",
              "title": "Project Name"
            },
            "description": "Directory name of the project"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchOptions"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_name}/search/replace-all": {
      "post": {
        "tags": ["Search"],
        "summary": "Replace All In Project",
        "description": "Replace every occurrence of a query throughout the active project.\n\nAll matching text in the specified scope is substituted with *replacement*.\nAfter completion the frontend should refresh story/chapter state.",
        "operationId": "replace_all_in_project_api_v1_projects__project_name__search_replace_all_post",
        "parameters": [
          {
            "name": "project_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Directory name of the project",
              "title": "Project Name"
            },
            "description": "Directory name of the project"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceAllRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplaceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_name}/search/replace-single": {
      "post": {
        "tags": ["Search"],
        "summary": "Replace Single In Project",
        "description": "Replace a single specifically identified match.\n\nThe match is identified by section_type, section_id, field, and match_index\n(zero-based ordinal of the match within that field).",
        "operationId": "replace_single_in_project_api_v1_projects__project_name__search_replace_single_post",
        "parameters": [
          {
            "name": "project_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Directory name of the project",
              "title": "Project Name"
            },
            "description": "Directory name of the project"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceSingleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplaceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReplaceAllRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query",
            "description": "Text or pattern to search for"
          },
          "replacement": {
            "type": "string",
            "title": "Replacement",
            "description": "Text to substitute in place of each match"
          },
          "scope": {
            "$ref": "#/components/schemas/SearchScope",
            "default": "all"
          },
          "case_sensitive": {
            "type": "boolean",
            "title": "Case Sensitive",
            "default": false
          },
          "is_regex": {
            "type": "boolean",
            "title": "Is Regex",
            "default": false
          },
          "is_phonetic": {
            "type": "boolean",
            "title": "Is Phonetic",
            "default": false
          },
          "active_chapter_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Chapter Id"
          }
        },
        "type": "object",
        "required": ["query", "replacement"],
        "title": "ReplaceAllRequest",
        "description": "Request to replace all occurrences of a search query."
      },
      "ReplaceResponse": {
        "properties": {
          "replacements_made": {
            "type": "integer",
            "title": "Replacements Made",
            "description": "Total number of occurrences replaced",
            "default": 0
          },
          "changed_sections": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Changed Sections",
            "description": "Human-readable labels for each changed section"
          }
        },
        "type": "object",
        "title": "ReplaceResponse",
        "description": "Result of a replace operation."
      },
      "ReplaceSingleRequest": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query",
            "description": "Text or pattern to search for"
          },
          "replacement": {
            "type": "string",
            "title": "Replacement",
            "description": "Text to substitute in place of the match"
          },
          "scope": {
            "$ref": "#/components/schemas/SearchScope",
            "default": "all"
          },
          "case_sensitive": {
            "type": "boolean",
            "title": "Case Sensitive",
            "default": false
          },
          "is_regex": {
            "type": "boolean",
            "title": "Is Regex",
            "default": false
          },
          "is_phonetic": {
            "type": "boolean",
            "title": "Is Phonetic",
            "default": false
          },
          "active_chapter_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Chapter Id"
          },
          "section_type": {
            "type": "string",
            "title": "Section Type",
            "description": "The section_type value from SearchResultSection"
          },
          "section_id": {
            "type": "string",
            "title": "Section Id",
            "description": "The section_id value from SearchResultSection"
          },
          "field": {
            "type": "string",
            "title": "Field",
            "description": "The field value from SearchResultSection"
          },
          "match_index": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Match Index",
            "description": "Zero-based index of the match within this section+field"
          }
        },
        "type": "object",
        "required": [
          "query",
          "replacement",
          "section_type",
          "section_id",
          "field",
          "match_index"
        ],
        "title": "ReplaceSingleRequest",
        "description": "Request to replace one specific match (identified by its ordinal index)."
      },
...
      "SearchOptions": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query",
            "description": "Text or pattern to search for"
          },
          "scope": {
            "$ref": "#/components/schemas/SearchScope",
            "description": "Which parts of the project to search",
            "default": "all"
          },
          "case_sensitive": {
            "type": "boolean",
            "title": "Case Sensitive",
            "description": "Whether the search is case-sensitive",
            "default": false
          },
          "is_regex": {
            "type": "boolean",
            "title": "Is Regex",
            "description": "Treat query as a regular expression",
            "default": false
          },
          "is_phonetic": {
            "type": "boolean",
            "title": "Is Phonetic",
            "description": "Use phonetic (soundex) matching instead of literal",
            "default": false
          },
          "active_chapter_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Chapter Id",
            "description": "Chapter ID to use when scope is current_chapter"
          }
        },
        "type": "object",
        "required": ["query"],
        "title": "SearchOptions",
        "description": "Parameters controlling what and how to search."
      },
      "SearchResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchResultSection"
            },
            "type": "array",
            "title": "Results"
          },
          "total_matches": {
            "type": "integer",
            "title": "Total Matches",
            "description": "Total number of individual matches",
            "default": 0
          }
        },
        "type": "object",
        "title": "SearchResponse",
        "description": "Top-level response for a search request."
      },
...
      "SearchScope": {
        "type": "string",
        "enum": ["current_chapter", "all_chapters", "sourcebook", "metadata", "all"],
        "title": "SearchScope",
        "description": "The scope of content to search within a project."
      },
      "SourcebookEntry": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "synonyms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Synonyms",
            "default": []
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "images": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Images",
            "default": []
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keywords",
            "default": []
          },
          "relations": {
            "items": {
              "$ref": "#/components/schemas/SourcebookRelation"
            },
            "type": "array",
            "title": "Relations",
            "default": []
          }
        },
        "type": "object",
        "required": ["id", "name", "description"],
        "title": "SourcebookEntry",
        "description": "Represents the SourcebookEntry type."
      },
      "SourcebookEntryCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "synonyms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Synonyms",
            "default": []
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "images": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Images",
            "default": []
          },
          "relations": {
            "items": {
              "$ref": "#/components/schemas/SourcebookRelation"
            },
            "type": "array",
            "title": "Relations",
            "default": []
          }
        },
        "type": "object",
        "required": ["name", "description"],
        "title": "SourcebookEntryCreate",
        "description": "Represents the SourcebookEntryCreate type."
      },
      "SourcebookEntryUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "synonyms": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Synonyms"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "images": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Images"
          },
          "relations": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SourcebookRelation"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relations"
          }
        },
        "type": "object",
        "title": "SourcebookEntryUpdate",
        "description": "Represents the SourcebookEntryUpdate type."
      },
      "SourcebookKeywordsRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "synonyms": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Synonyms"
          }
        },
        "type": "object",
        "title": "SourcebookKeywordsRequest",
        "description": "Request payload for generating keywords from an entry description."
      },
      "SourcebookKeywordsResponse": {
        "properties": {
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keywords"
          }
        },
        "type": "object",
        "required": ["keywords"],
        "title": "SourcebookKeywordsResponse",
        "description": "Represents the SourcebookKeywordsResponse type."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}